refactor: remove unused ComicDetailPage component and update RelatedComics to use nftAddress instead of id
